fix type mismatch

+++ structGenAPI/src/main/java/net/piratjsk/structgen/loaders/structfiles/ConditionFactory.java
@@ -15,7 +15,7 @@ public class ConditionFactory {
                 final int y = data.getInt("pos.y");
                 final int z = data.getInt("pos.z");
                 final int id = data.getInt("id");
-                final int meta = data.hasPath("meta") ? data.getInt("meta") : 0;
+                final byte meta = data.hasPath("meta") ? (byte) data.getInt("meta") : 0;
                 return new BlockCondition(x, y, z, id, meta);
             case "conditionGroup":
                 final Collection<Condition> conditions = new ArrayList<>();
